asm: refactor type resolution code
