+++ asm/helper.go
@@ -1,0 +1,73 @@
+package asm
+
+import (
+	"fmt"
+	"strconv"
+	"strings"
+
+	"github.com/llir/l/ir/types"
+	"github.com/mewmew/l-tm/asm/ll/ast"
+	"github.com/mewmew/l-tm/internal/enc"
+)
+
+// irAddrSpace returns the IR address space corresponding to the given optional
+// AST address space.
+func irAddrSpace(n *ast.AddrSpace) types.AddrSpace {
+	// \empty is used when address space not present.
+	if n.Text() == "" {
+		return 0
+	}
+	x := uintLit(n.N())
+	return types.AddrSpace(x)
+}
+
+// local returns the name (without '%' prefix) of the given local identifier.
+func local(l ast.LocalIdent) string {
+	text := l.Text()
+	const prefix = "%"
+	if !strings.HasPrefix(text, prefix) {
+		// NOTE: Panic instead of returning error as this case should not be
+		// possible given the grammar.
+		panic(fmt.Errorf("invalid local identifier %q; missing '%s' prefix", text, prefix))
+	}
+	text = text[len(prefix):]
+	return unquote(text)
+}
+
+// uintLit returns the unsigned integer value corresponding to the given
+// unsigned integer literal.
+func uintLit(l ast.UintLit) uint64 {
+	text := l.Text()
+	x, err := strconv.ParseUint(text, 10, 64)
+	if err != nil {
+		// NOTE: Panic instead of returning error as this case should not be
+		// possible given the grammar.
+
+		// TODO: figure out how to update the grammar for UintLit to remove the
+		// optional sign.
+		panic(fmt.Errorf("unable to parse unsigned integer literal %q; %v", text, err))
+	}
+	return x
+}
+
+// unquote returns the unquoted version of s if quoted, and the original string
+// otherwise.
+func unquote(s string) string {
+	if len(s) >= 2 && strings.HasPrefix(s, `"`) && strings.HasSuffix(s, `"`) {
+		return string(enc.Unquote(s))
+	}
+	return s
+}
+
+// irVariadic returns the variadic boolean corresponding to the given optional
+// AST ellipsis.
+func irVariadic(n *ast.Ellipsis) bool {
+	// TODO: check why Variadic is non-nil for `Variadic=Ellipsisopt`, regardless
+	// of whether the input is (...) or ().
+	//
+	// It seems that the Variadic.Text simply returns empty string when
+	// Ellipsisopt reduces to \empty.
+	//
+	// Using `n.Text() == "..."` for now, would like to use `n != nil`.
+	return n.Text() == "..."
+}
